% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_i.R
\name{effective_i}
\alias{effective_i}
\title{Calculate effective annual interest rate for a known nominal rate and
compounding period per year

\code{effective_i} Compute future value nominal rate and
compounding period per year, and frequency.}
\usage{
effective_i(r, m = c("yearly", "quarterly", "monthly"))
}
\arguments{
\item{r}{nominal interest rate in decimal number}

\item{m}{number of compounding period per year
monthly = 12
Quarterly = 4
yearly = 1}
}
\value{
effective_i
}
\description{
Calculate effective annual interest rate for a known nominal rate and
compounding period per year

\code{effective_i} Compute future value nominal rate and
compounding period per year, and frequency.
}
\examples{
 
# Result: effective_i = 
effective_i (0.08, m = "yearly")  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y. (2020).
Engineering Economic Analysis (14th ed.).
New York, Oxford University Press.

David, W., & Terry, R. (2012).
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature.
}
