% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_to_annual.R
\name{future_to_annual}
\alias{future_to_annual}
\title{Compute annual payment value from future value

\code{future_to_annual} calculate annual value from future value using
the accumulated amount after years using sinking fund factor (sff)}
\usage{
future_to_annual(i, n, FV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{FV}{accumulated (future) value}
}
\value{
A
}
\description{
Compute annual payment value from future value

\code{future_to_annual} calculate annual value from future value using
the accumulated amount after years using sinking fund factor (sff)
}
\examples{
 
# Result: A = 5737.83
future_to_annual(0.08, 30, 650000)  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
