% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_to_present.R
\name{future_to_present}
\alias{future_to_present}
\title{Compute present values from future payments (or cash flow)

\code{future_to_present} calculate present value from future value using
present worth factor (pwf)}
\usage{
future_to_present(i, n, FV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{FV}{future value}
}
\value{
pr.value
}
\description{
Compute present values from future payments (or cash flow)

\code{future_to_present} calculate present value from future value using
present worth factor (pwf)
}
\examples{
 
# Result: pr.value = 2981.32
future_to_present(0.08, 30, 30000)  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
