% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_to_annual.R
\name{gradient_to_annual}
\alias{gradient_to_annual}
\title{Calculate annual value from uniform gradient payment

\code{gradient_to_annual} Compute annual value from uniform gradient payments
using uniform gra present worth factor (ugaw)}
\usage{
gradient_to_annual(i, n, G, A)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}

\item{A}{initial annual payment}
}
\value{
ann.value
}
\description{
Calculate annual value from uniform gradient payment

\code{gradient_to_annual} Compute annual value from uniform gradient payments
using uniform gra present worth factor (ugaw)
}
\examples{
 
# Result: ann.value = $ 105138.30
gradient_to_annual(0.08, 30, 6000, 50000)  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
