% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_to_future.R
\name{gradient_to_future}
\alias{gradient_to_future}
\title{Calculate future value from uniform gradient payment

\code{gradient_to_future} Compute future value from uniform gradient payments
using uniform gradient present worth factor (ugfw)}
\usage{
gradient_to_future(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
fr.value
}
\description{
Calculate future value from uniform gradient payment

\code{gradient_to_future} Compute future value from uniform gradient payments
using uniform gradient present worth factor (ugfw)
}
\examples{
 
# Result: fr.value = $ 312312
gradient_to_future(0.08, 30,300)  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
