% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_to_present.R
\name{gradient_to_present}
\alias{gradient_to_present}
\title{Calculate present value from uniform gradient payment

\code{gradient_to_present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)}
\usage{
gradient_to_present(i, n, G)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{G}{uniform gradient payments}
}
\value{
pr.value
}
\description{
Calculate present value from uniform gradient payment

\code{gradient_to_present} Compute present value from uniform gradient payments
using uniform gradient present worth factor (ugpw)
}
\examples{
 
# Result: pr.value = 310367.40
gradient_to_present(0.08, 30, 3000)

}
\references{
#'  Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
