% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_during_construction.R
\name{interest_during_construction}
\alias{interest_during_construction}
\title{Compute interest during construction

\code{interest_during_construction} calculate interest during construction
using a monthly discount factor}
\usage{
interest_during_construction(i, duration, capital)
}
\arguments{
\item{i}{interest rate in percent per year}

\item{duration}{construction duration in months}

\item{capital}{first/capital cost}
}
\value{
idc
}
\description{
Compute interest during construction

\code{interest_during_construction} calculate interest during construction
using a monthly discount factor
}
\examples{
 
# Result: idc ($) = 18992.14
interest_during_construction (0.027, 25, 700000)

}
\references{
Engineer, U. S. A., & Resources, W.
National Economic Development (NED) Procedures Manual- National Development Costs.
U.S. Army Corps of Engineers Humphreys Engineer Center Support
Activity Institute for Water Resources, DACWC72-90(June)1993.
}
