% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_distribute.R
\name{om_distribute}
\alias{om_distribute}
\title{Calculate the present value of periodic operations and maintenance costs

\code{om_distribute} Distribute periodic present value operations and maintenance costs
over a project life span, discount over project, and compute present value}
\usage{
om_distribute(i, n, fq, OM)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{fq}{frequency of cost in years}

\item{OM}{operation and maintenance cost incurred at each interval in present value}
}
\value{
OM_dist
}
\description{
Calculate the present value of periodic operations and maintenance costs

\code{om_distribute} Distribute periodic present value operations and maintenance costs
over a project life span, discount over project, and compute present value
}
\examples{

#Result is the present value cost of periodic operations and maintenance expenses.
# Result: idc ($) = 8174.547
om_distribute(0.03, 50, 3, 1000)  

}
\references{
Add citation as needed.
}
