% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_to_future.R
\name{present_to_future}
\alias{present_to_future}
\title{Compute future values from present value}
\usage{
present_to_future(i, n, PV)
}
\arguments{
\item{i}{discount rate in percent per year}

\item{n}{life span in years}

\item{PV}{present value}
}
\value{
fr.value
}
\description{
\code{present_to_future} calculate future value from present value using
compound amount factor (caf)
}
\examples{
 
# Result: fr.value = 9056391

present_to_future (0.08, 30, 900000)  

}
\references{
Newnan, D. G., Eschenbach, T. G., Lavelle, J. P., & Oxford, N. Y.
Engineering Economic Analysis, 14th ed.
New York, Oxford University Press, 2020

David, W., & Terry, R.
Fundamentals of Engineering Economics and Decision Analysis.
Springer Nature, 2012
}
